+++ packages/in_app_purchase/in_app_purchase_android/test/in_app_purchase_android_platform_addition_test.dart
@@ -39,7 +39,7 @@ void main() {
         value: buildBillingResultMap(expectedBillingResult));
     stubPlatform.addResponse(name: endConnectionCall);
     iapAndroidPlatformAddition =
-        InAppPurchaseAndroidPlatformAddition(BillingClient((_) {}));
+        InAppPurchaseAndroidPlatformAddition(BillingClientManager());
   });
 
   group('consume purchases', () {